Allow direct assignment of results
As of Ruby 2.2 (I went back and tested successive versions to find out),
it's possible to assign the result of the http request directly. In 2.1
and below that was a syntax error, so we had to do all of the weird
`binding_of_caller` nonsense to assign the result.

This commit gets rid of that, and brings us back to the original vision
of this terrible idea.